# SpeakerDesign/api/v1/openapi.yaml
openapi: 3.0.3
info:
  title: Loudspeaker Foundation API
  description: |
    REST API for loudspeaker enclosure design based on Thiele-Small parameters.

    **Built on pure theory from published papers:**
    - Small 1972: Sealed box calculations
    - Thiele 1971: Standard alignments
    - Small 1973: Ported box calculations

    All calculations cite source equations. Accepts both SI units (m³) and user-friendly units (liters).

  version: 0.1.0
  contact:
    name: Foundation Library
    url: https://github.com/YOUR_USERNAME/loudspeaker-foundation
  license:
    name: MIT
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
...
paths:
...
  /validate/driver:
    post:
      tags: [Validation]
      operationId: validateDriver
      summary: Validate Thiele-Small parameters
      description: |
        Validate driver T/S parameters against theoretical limits and practical ranges.

        Valid ranges:
        - Fs: 15-500 Hz (pistonic behavior)
        - Qts: 0.2-1.5 (practical damping)
        - Vas: > 0
        - Qes: > Qts (by definition)

        Source: Small 1972 + Dickason 2006
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DriverValidationRequest'
      responses:
        '200':
          description: Validation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationResponse'
...
components:
  schemas:
    Driver:
      type: object
      required: [fs, qts, vas]
      properties:
        fs:
          type: number
          description: Free-air resonance frequency
          minimum: 15
          maximum: 500
          example: 22.0
        qts:
          type: number
          description: Total Q factor
          minimum: 0.2
          maximum: 1.5
          example: 0.53
        qes:
          type: number
          description: Electrical Q factor (optional)
          minimum: 0.2
          maximum: 2.0
          example: 0.56
        vas:
          type: number
          description: Equivalent compliance volume
          minimum: 0.001
          example: 248.2
        vasUnit:
          type: string
          enum: [m3, L, liters, cuft]
          default: m3
          description: Unit for Vas (defaults to m³)
...
    DriverValidationRequest:
      type: object
      required: [driver]
      properties:
        driver:
          $ref: '#/components/schemas/Driver'
    ValidationResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            valid:
              type: boolean
            warnings:
              type: array
              items:
                type: string
            errors:
              type: array
              items:
                type: string
...